pyproject.toml
[tool.poetry]
name = "example-fastapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.23"
pydantic-settings = "^2.1.0"
psycopg = {extras = ["binary"], version = "^3.1.15"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
email-validator = "^2.1.0.post1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
